+++ watchme/tests/test_decorators.py
@@ -11,8 +11,6 @@ from watchme.command import create_watcher
 import unittest
 import tempfile
 import shutil
-import json
-import os
 
 
 print("########################################################### test_client")