+++ watchme/utils/__init__.py
@@ -1,7 +1,4 @@
 from .terminal import (
-    confirm_prompt,
-    choice_prompt,
-    regexp_prompt,
     run_command,
     get_installdir,
     get_watchme_env,